Added data rest

+++ src/main/java/com/example/opbook/model/Post.java
@@ -9,12 +9,12 @@ import java.util.Set;
 @Entity
 public class Post {
     @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "ID")
-    private Integer id;
+    private Long id;
 
     @ManyToOne
-    @JoinColumn(name="SubmitterID", nullable=false)
+    @JoinColumn(name = "SubmitterID", nullable = false)
     private User submitter;
 
     @Column(name = "CreationTime")
@@ -30,7 +30,7 @@ public class Post {
     private String content;
 
     @ManyToOne
-    @JoinColumn(name = "CourseID", nullable=false)
+    @JoinColumn(name = "CourseID", nullable = false)
     private Course course;
 
     @ManyToMany(mappedBy = "likedPosts")
@@ -39,33 +39,55 @@ public class Post {
     @OneToMany(mappedBy = "post")
     private Set<PostComment> comments;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public User getSubmitter() { return submitter; }
+    public User getSubmitter() {
+        return submitter;
+    }
 
-    public void setSubmitter(User submitter) { this.submitter = submitter; }
+    public void setSubmitter(User submitter) {
+        this.submitter = submitter;
+    }
 
-    public Date getCreationTime() { return creationTime; }
+    public Date getCreationTime() {
+        return creationTime;
+    }
 
-    public void setCreationTime(Date creationTime) { this.creationTime = creationTime; }
+    public void setCreationTime(Date creationTime) {
+        this.creationTime = creationTime;
+    }
 
-    public String getTitle() { return title; }
+    public String getTitle() {
+        return title;
+    }
 
-    public void setTitle(String title) { this.title = title; }
+    public void setTitle(String title) {
+        this.title = title;
+    }
 
-    public String getContent() { return content; }
+    public String getContent() {
+        return content;
+    }
 
-    public void setContent(String content) { this.content = content;}
+    public void setContent(String content) {
+        this.content = content;
+    }
 
-    public Course getCourse() { return course; }
+    public Course getCourse() {
+        return course;
+    }
 
-    public void setCourse(Course course) { this.course = course;}
+    public void setCourse(Course course) {
+        this.course = course;
+    }
 
-    public Set<PostComment> getComments() { return comments; }
+    public Set<PostComment> getComments() {
+        return comments;
+    }
 }
